chore: move rwmap and syncmap to maps pkg
